添加gson和接口文件

+++ XunPictureTranslate/src/main/java/com/xxun/watch/picture_translate/bean/IdentityPictureResponse.kt
@@ -11,7 +11,7 @@ data class DataArray(
 )
 
 data class Word(
-    val conf: Double,
+    val conf: Float,
     val content: String,
     val coord: List<Coord>
 )
